Fix ruff RUF046 lint failures blocking CI
- execution/broker_alpaca.py: redundant int() cast introduced by the
  fractional-short-order fix (round() on a float already returns an int)
- execution/broker_ibkr.py: same redundant cast, pre-existing on main —
  this is why CI has been failing on every push regardless of PR content

+++ execution/broker_alpaca.py
@@ -92,7 +92,7 @@ class AlpacaBroker:
         qty = abs(round(delta, 4))
 
         if side == OrderSide.SELL and target_shares < 0:
-            qty = float(int(round(qty)))
+            qty = float(round(qty))
             if qty < 1:
                 logger.info("Skipping %s: rounds to 0 whole shares for a short order.", symbol)
                 return None
